+++ src/main/java/racingcar/domain/car/CarDto.java
@@ -1,4 +1,4 @@
-package racingcar.dto;
+package racingcar.domain.car;
 
 public record CarDto(String name, int position) {
 }
