+++ src/main/java/racingcar/domain/game/Round.java
@@ -3,7 +3,7 @@ package racingcar.domain.game;
 import java.util.List;
 import racingcar.domain.car.Car;
 import racingcar.domain.car.Cars;
-import racingcar.dto.CarDto;
+import racingcar.domain.car.CarDto;
 import racingcar.mapper.CarMapper;
 
 public class Round {
