+++ src/main/java/racingcar/view/ConsoleOutputView.java
@@ -3,7 +3,7 @@ package racingcar.view;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import racingcar.dto.CarDto;
+import racingcar.domain.car.CarDto;
 
 public class ConsoleOutputView implements OutputView{
     private static final String REQUEST_CAR_NAMES_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
